refactor(apps/frontend-manage): simplify structure of live quiz activity list entries in preparation for generalization
